Updated structure and created new luckperms version
-Updated to newer layout for the source code and Maven versions.
-Updated libraries.
-Fixed issue with base version where configuration file keeps resetting on newer versions.
-Added save config on each time a player sets a claim to avoid loss after crash.
-Added Remove Method to be used in unclaim process that interfaces better with the config file.
-Added new LuckPerms version that integrates with LuckPerms. Groups will now automatically be generated in clean configs.
Group names are now LuckPerms group names. Each group from LuckPerms can now have its own set of maxclaimedblocks.
