+++ .gitignore
@@ -1,0 +1,4 @@
+/target/
+.project
+/.settings/
+.classpath
