+++ LandClaiming - LuckPerms/src/main/java/me/zombie_striker/landclaiming/claimedobjects/ClaimedBlock.java
