+++ LandClaiming - LuckPerms/src/main/java/me/zombie_striker/landclaiming/commands/LockCommand.java
