+++ LandClaiming - base/pom.xml
@@ -3,53 +3,43 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
     <groupId>me.zombie_striker</groupId>
-    <artifactId>landclaiming</artifactId>
+    <artifactId>landclaiming-base</artifactId>
     <version>1.2.2</version>
+    <packaging>jar</packaging>
+    <name>LandClaiming</name>
     <build>
+		<defaultGoal>clean package</defaultGoal>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.11.0</version>
                 <configuration>
-                    <source>8</source>
-                    <target>8</target>
+				<source>8</source>
+                <target>8</target>
                 </configuration>
             </plugin>
         </plugins>
     </build>
-
     <repositories>
         <repository>
             <id>spigot-repo</id>
             <url>https://hub.spigotmc.org/nexus/content/repositories/snapshots/</url>
         </repository>
-        <repository>
-			<id>EssentialsEx</id>
-			<url>https://repo.essentialsx.net/#/releases/net/essentialsx/EssentialsX/2.20.1</url>
-		</repository>
     </repositories>
-
     <dependencies>
         <dependency>
             <groupId>org.spigotmc</groupId>
             <artifactId>spigot-api</artifactId>
-            <version>1.20.2-R0.1-SNAPSHOT</version>
+            <version>1.18-R0.1-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
-        <!-- https://mvnrepository.com/artifact/com.googlecode.json-simple/json-simple -->
         <dependency>
             <groupId>com.googlecode.json-simple</groupId>
             <artifactId>json-simple</artifactId>
             <version>1.1.1</version>
         </dependency>
-        <dependency>
-        	<groupId>net.luckperms</groupId>
-        	<artifactId>api</artifactId>
-        	<version>5.4</version>
-        	<scope>provided</scope>
-    	</dependency>
     </dependencies>
 
 </project>
