--- plugin.yml
@@ -1,25 +1,0 @@
-main: me.zombie_striker.landclaiming.LandClaiming
-version: 1.4.11
-name: LandClaiming
-api-version: 1.13
-commands:
-   claim:
-      description: Claims land
-   unclaim:
-      description: Unclaims land
-   lock:
-      description: Locks blocks
-   unlock:
-      description: Unlocks blocks
-author:
-   Zombie_Striker
-permissions:
-   landclaim.*:
-     description: Gives access to all Inferno commands
-     children:
-       landclaim.default: true
-       landclaim.admin: true
-   landclaim.admin:
-     description: Gives access to all Inferno commands
-     children:
-       landclaim.default: true
