--- src/module-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * 
- */
-module landclaiming_new {
-}
